fix: options form errors
